+++ sdk/resourcemanager/peering/armpeering/ze_generated_example_connectionmonitortests_client_test.go
@@ -17,79 +17,81 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/peering/armpeering"
 )
 
-// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/GetPeeringServicePrefix.json
-func ExamplePrefixesClient_Get() {
+// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/GetConnectionMonitorTest.json
+func ExampleConnectionMonitorTestsClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client := armpeering.NewPrefixesClient("<subscription-id>", cred, nil)
+	client := armpeering.NewConnectionMonitorTestsClient("<subscription-id>", cred, nil)
 	res, err := client.Get(ctx,
 		"<resource-group-name>",
 		"<peering-service-name>",
-		"<prefix-name>",
-		&armpeering.PrefixesClientGetOptions{Expand: nil})
+		"<connection-monitor-test-name>",
+		nil)
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.Printf("Response result: %#v\n", res.PrefixesClientGetResult)
+	log.Printf("Response result: %#v\n", res.ConnectionMonitorTestsClientGetResult)
 }
 
-// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/CreatePeeringServicePrefix.json
-func ExamplePrefixesClient_CreateOrUpdate() {
+// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/CreateOrUpdateConnectionMonitorTest.json
+func ExampleConnectionMonitorTestsClient_CreateOrUpdate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client := armpeering.NewPrefixesClient("<subscription-id>", cred, nil)
+	client := armpeering.NewConnectionMonitorTestsClient("<subscription-id>", cred, nil)
 	res, err := client.CreateOrUpdate(ctx,
 		"<resource-group-name>",
 		"<peering-service-name>",
-		"<prefix-name>",
-		armpeering.ServicePrefix{
-			Properties: &armpeering.ServicePrefixProperties{
-				PeeringServicePrefixKey: to.StringPtr("<peering-service-prefix-key>"),
-				Prefix:                  to.StringPtr("<prefix>"),
+		"<connection-monitor-test-name>",
+		armpeering.ConnectionMonitorTest{
+			Properties: &armpeering.ConnectionMonitorTestProperties{
+				Destination:        to.StringPtr("<destination>"),
+				DestinationPort:    to.Int32Ptr(443),
+				SourceAgent:        to.StringPtr("<source-agent>"),
+				TestFrequencyInSec: to.Int32Ptr(30),
 			},
 		},
 		nil)
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.Printf("Response result: %#v\n", res.PrefixesClientCreateOrUpdateResult)
+	log.Printf("Response result: %#v\n", res.ConnectionMonitorTestsClientCreateOrUpdateResult)
 }
 
-// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/DeletePeeringServicePrefix.json
-func ExamplePrefixesClient_Delete() {
+// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/DeleteConnectionMonitorTest.json
+func ExampleConnectionMonitorTestsClient_Delete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client := armpeering.NewPrefixesClient("<subscription-id>", cred, nil)
+	client := armpeering.NewConnectionMonitorTestsClient("<subscription-id>", cred, nil)
 	_, err = client.Delete(ctx,
 		"<resource-group-name>",
 		"<peering-service-name>",
-		"<prefix-name>",
+		"<connection-monitor-test-name>",
 		nil)
 	if err != nil {
 		log.Fatal(err)
 	}
 }
 
-// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/ListPrefixesByPeeringService.json
-func ExamplePrefixesClient_ListByPeeringService() {
+// x-ms-original-file: specification/peering/resource-manager/Microsoft.Peering/stable/2021-06-01/examples/ListConnectionMonitorTestsByPeeringService.json
+func ExampleConnectionMonitorTestsClient_ListByPeeringService() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client := armpeering.NewPrefixesClient("<subscription-id>", cred, nil)
+	client := armpeering.NewConnectionMonitorTestsClient("<subscription-id>", cred, nil)
 	pager := client.ListByPeeringService("<resource-group-name>",
 		"<peering-service-name>",
-		&armpeering.PrefixesClientListByPeeringServiceOptions{Expand: nil})
+		nil)
 	for {
 		nextResult := pager.NextPage(ctx)
 		if err := pager.Err(); err != nil {